LayoutTests/media/modern-media-controls/scrubber/scrubber-expected.txt

Testing the Scrubber class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS scrubber.element.classList.contains("scrubber") is true
PASS successfullyParsed is true

TEST COMPLETE
